Moved IdGenerator to common port folder as it used in recipe and user.
