+++ src/main/java/recipe/foodbar/app/manual/TestFilterRecipes.java
@@ -4,7 +4,7 @@ import com.mongodb.client.MongoDatabase;
 import recipe.foodbar.entities.*;
 import recipe.foodbar.id_generator.jug.JugIdGenerator;
 import recipe.foodbar.repository.mongoDB.repository.MongoRecipeRepository;
-import recipe.foodbar.usecase.user.port.IdGenerator;
+import recipe.foodbar.usecase.commonport.IdGenerator;
 
 import java.util.ArrayList;
 import java.util.Date;
