+++ src/main/java/recipe/foodbar/usecase/cuisine/CreateCuisine.java
@@ -4,7 +4,7 @@ import recipe.foodbar.entities.Cuisine;
 import recipe.foodbar.usecase.cuisine.exception.CuisineAlreadyExistsException;
 import recipe.foodbar.usecase.cuisine.port.CuisineRepository;
 import recipe.foodbar.usecase.cuisine.validator.CuisineValidator;
-import recipe.foodbar.usecase.user.port.IdGenerator;
+import recipe.foodbar.usecase.commonport.IdGenerator;
 
 public class CreateCuisine {
     private final CuisineRepository repository;
