+++ src/main/java/recipe/foodbar/usecase/recipe/manager/CreateRecipe.java
@@ -4,7 +4,7 @@ import recipe.foodbar.entities.Recipe;
 import recipe.foodbar.usecase.recipe.exception.RecipeAlreadyExistsException;
 import recipe.foodbar.usecase.recipe.port.RecipeRepository;
 import recipe.foodbar.usecase.recipe.validator.RecipeValidator;
-import recipe.foodbar.usecase.user.port.IdGenerator;
+import recipe.foodbar.usecase.commonport.IdGenerator;
 
 public class CreateRecipe {
     private final RecipeRepository repository;
