+++ src/main/java/recipe/foodbar/usecase/user/UserManager.java
@@ -5,7 +5,7 @@
 package recipe.foodbar.usecase.user;
 
 
-import recipe.foodbar.usecase.user.port.IdGenerator;
+import recipe.foodbar.usecase.commonport.IdGenerator;
 import recipe.foodbar.usecase.user.port.UserCreatorInputBoundary;
 import recipe.foodbar.usecase.user.port.UserCreatorOutputBoundary;
 import recipe.foodbar.usecase.user.port.UserRepositoryInterface;
