--- src/main/java/recipe/foodbar/usecase/user/port/IdGenerator.java
@@ -1,5 +1,0 @@
-package recipe.foodbar.usecase.user.port;
-
-public interface IdGenerator {
-    String generate();
-}
