+++ src/test/java/UserTest.java
@@ -6,7 +6,7 @@ import recipe.foodbar.id_generator.jug.JugIdGenerator;
 import recipe.foodbar.repository.mongoDB.repository.MongoUserRepository;
 import recipe.foodbar.usecase.user.UserInputData;
 import recipe.foodbar.usecase.user.UserManager;
-import recipe.foodbar.usecase.user.port.IdGenerator;
+import recipe.foodbar.usecase.commonport.IdGenerator;
 import recipe.foodbar.usecase.user.port.UserCreatorInputBoundary;
 
 import static recipe.foodbar.repository.mongoDB.MongoDB.getMongoDB;
